+++ packages/react-native-eyecandy-theme/src/colors/createColors.test.ts
@@ -103,6 +103,7 @@ test('createColors', () => {
     primary: badgeColor,
     success: badgeColor,
     warning: badgeColor,
+    danger: badgeColor,
   });
 });