+++ packages/react-native-eyecandy/src/components/SearchBar/SearchBar.tsx
@@ -4,7 +4,7 @@ import { StyleSheet, View } from 'react-native';
 import { Search, Filter } from '@nomada-sh/react-native-eyecandy-icons';
 
 import { Button } from '../Button';
-import TextInput, { TextInputProps } from '../TextInput';
+import { TextInput, TextInputProps } from '../TextInput';
 
 export interface SearchBarProps
   extends Omit<TextInputProps, 'onPressIconRight'> {