+++ packages/react-native-eyecandy/src/components/TextInputPassword/TextInputPassword.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Lock } from '@nomada-sh/react-native-eyecandy-icons';
 
-import TextInput, { TextInputProps } from '../TextInput';
+import { TextInput, TextInputProps } from '../TextInput';
 
 function TextInputPassword(props: TextInputProps) {
   return (