+++ packages/react-native-eyecandy/src/components/TextInputUsername/TextInputUsername.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { User } from '@nomada-sh/react-native-eyecandy-icons';
 
-import TextInput, { TextInputProps } from '../TextInput';
+import { TextInput, TextInputProps } from '../TextInput';
 
 function TextInputUsername(props: TextInputProps) {
   return (