+++ packages/react-native-eyecandy/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nomada-sh/react-native-eyecandy
 
+## 0.0.12
+
+### Patch Changes
+
+- Added missing TextInputHandle export from TextInput component.
+- Updated dependencies
+  - @nomada-sh/react-native-eyecandy-theme@0.0.5
+  - @nomada-sh/react-native-eyecandy-icons@0.0.7
+
 ## 0.0.11
 
 ### Patch Changes